requirements.txt
# TransWacom Requirements
# Core dependencies for the modular TransWacom system

# Input/output handling
evdev>=1.3.0

# Network discovery
zeroconf>=0.38.0

# Configuration management
pyyaml>=6.0

# GUI dependencies
pystray>=0.19.0
Pillow>=9.0.0
plyer>=2.1.0  # For desktop notifications
PyGObject>=3.42.0  # Required for pystray on Linux
